+++ src/com/oltpbenchmark/benchmarks/wikipedia/procedures/RemoveWatchList.java
@@ -38,11 +38,11 @@ public class RemoveWatchList extends Procedure {
         " WHERE user_id =  ? "
     ); 
 
-    public void run(Connection conn, int userId, int nameSpace, String pageTitle) throws SQLException {
+    public void run(Connection conn, long userId, int nameSpace, String pageTitle) throws SQLException {
 
         if (userId > 0) {
             PreparedStatement ps = this.getPreparedStatement(conn, removeWatchList);
-            ps.setInt(1, userId);
+            ps.setLong(1, userId);
             ps.setInt(2, nameSpace);
             ps.setString(3, pageTitle);
             ps.executeUpdate();
@@ -51,7 +51,7 @@ public class RemoveWatchList extends Procedure {
                 // if regular page, also remove a line of
                 // watchlist for the corresponding talk page
                 ps = this.getPreparedStatement(conn, removeWatchList);
-                ps.setInt(1, userId);
+                ps.setLong(1, userId);
                 ps.setInt(2, 1);
                 ps.setString(3, pageTitle);
                 ps.executeUpdate();
@@ -59,7 +59,7 @@ public class RemoveWatchList extends Procedure {
 
             ps = this.getPreparedStatement(conn, setUserTouched);
             ps.setString(1, TimeUtil.getCurrentTimeString14());
-            ps.setInt(2, userId);
+            ps.setLong(2, userId);
             ps.executeUpdate();
         }
     }
